relocation methods for time in service
